+++ src/main/java/com/climatizacion/sistema_clima/service/CitaService.java
@@ -10,8 +10,8 @@ public interface CitaService {
     List<CitaResponseDTO> obtenerTodas();
     List<CitaResponseDTO> obtenerPorTecnico(Long idTecnico);
     CitaResponseDTO crear(CitaRequestDTO request);
-    CitaResponseDTO actualizar(Integer id, CitaRequestDTO request);
-    void cambiarEstado(Integer id, String estado);
+    CitaResponseDTO actualizar(Long id, CitaRequestDTO request);
+    void cambiarEstado(Long id, String estado);
     List<CitaResponseDTO> obtenerPorCliente(Long idCliente);
     long contarCitasPorClienteYEstados(Long idCliente, List<EstadoCita> estados);
 }
