+++ src/main/java/com/climatizacion/sistema_clima/service/DetalleCitaService.java
@@ -5,7 +5,7 @@ import com.climatizacion.sistema_clima.dto.DetalleCitaResponseDTO;
 import java.util.List;
 
 public interface DetalleCitaService {
-    List<DetalleCitaResponseDTO> obtenerPorCita(Integer idCita);
+    List<DetalleCitaResponseDTO> obtenerPorCita(Long idCita);
     DetalleCitaResponseDTO crear(DetalleCitaRequestDTO request);
-    void eliminar(Integer idDetalleCita);
+    void eliminar(Long idDetalleCita);
 }
